fix: resolve pr comments
