+++ src/app/pages/SpotTradingPage/components/SpotHistory/AssetRow.tsx
@@ -5,7 +5,7 @@ import { LoadableValue } from 'app/components/LoadableValue';
 import { useGetProfitDollarValue } from 'app/hooks/trading/useGetProfitDollarValue';
 import { getOrder, TradingTypes } from 'app/pages/SpotTradingPage/types';
 import { translations } from 'locales/i18n';
-import React from 'react';
+import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { TxStatus } from 'store/global/transactions-store/types';
 import { toNumberFormat } from 'utils/display-text/format';
@@ -14,13 +14,14 @@ import iconSuccess from 'assets/images/icon-success.svg';
 import iconRejected from 'assets/images/icon-rejected.svg';
 import iconPending from 'assets/images/icon-pending.svg';
 import { AssetValue } from 'app/components/AssetValue';
+import { Nullable } from 'types';
 
 interface IAssetRowProps {
   status?: TxStatus;
   timestamp: number;
   transactionHash: string;
   fromAmount: string;
-  toAmount: string | null;
+  toAmount: Nullable<string>;
   assetFrom: AssetDetails;
   assetTo: AssetDetails;
 }
@@ -41,7 +42,10 @@ export const AssetRow: React.FC<IAssetRowProps> = ({
     toAmount || '0',
   );
 
-  const order = getOrder(assetFrom.asset, assetTo.asset);
+  const order = useMemo(() => getOrder(assetFrom.asset, assetTo.asset), [
+    assetFrom,
+    assetTo,
+  ]);
 
   if (!order) {
     return null;
@@ -70,8 +74,11 @@ export const AssetRow: React.FC<IAssetRowProps> = ({
         </span>
       </td>
       <td>
-        {toNumberFormat(fromAmount, 4)}{' '}
-        <AssetRenderer asset={assetFrom.asset} />
+        <AssetValue
+          value={Number(fromAmount)}
+          minDecimals={4}
+          asset={assetFrom.asset}
+        />
       </td>
       <td className="tw-hidden lg:tw-table-cell">
         <AssetValue value={Number(toAmount || 0)} asset={assetTo.asset} />
